pyproject.toml
[project]
name = "elabftw-opencloning-sync"
version = "0.1.0"
description = "Sync eLabFTW sequences to OpenCloning local static content"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "python-dotenv",
    "requests"
]